+++ app/src/main/java/woowacourse/movie/BookingActivity.kt
@@ -4,12 +4,12 @@ import android.content.Intent
 import android.os.Bundle
 import android.view.View
 import android.widget.AdapterView
-import android.widget.ArrayAdapter
 import android.widget.Button
 import android.widget.ImageView
 import android.widget.Spinner
 import android.widget.TextView
 import androidx.appcompat.app.AppCompatActivity
+import woowacourse.movie.adapter.SpinnerAdapter
 import woowacourse.movie.data.MovieInfo
 import woowacourse.movie.data.Ticket
 import java.time.DayOfWeek
