+++ app/src/main/java/woowacourse/movie/adapter/MovieListAdapter.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie
+package woowacourse.movie.adapter
 
 import android.content.Context
 import android.content.Intent
@@ -9,6 +9,8 @@ import android.widget.ArrayAdapter
 import android.widget.Button
 import android.widget.ImageView
 import android.widget.TextView
+import woowacourse.movie.BookingActivity
+import woowacourse.movie.R
 import woowacourse.movie.data.MovieInfo
 
 class MovieListAdapter(context: Context, items: MutableList<MovieInfo>): ArrayAdapter<MovieInfo>(context, 0, items) {
@@ -33,16 +35,8 @@ class MovieListAdapter(context: Context, items: MutableList<MovieInfo>): ArrayAd
         val button = view.findViewById<Button>(R.id.reservation_button)
         button.setOnClickListener {
             if (item != null) {
-                val movieInfo = MovieInfo(
-                    poster = item.poster,
-                    title = title.text.toString(),
-                    startDate = startDate.text.toString(),
-                    endDate = endDate.text.toString(),
-                    runningTime = runningTime.text.toString()
-                )
-
                 val intent = Intent(context, BookingActivity::class.java).apply {
-                    putExtra("MOVIE_INFO", movieInfo)
+                    putExtra("MOVIE_INFO", item)
                 }
                 context.startActivity(intent)
             }
